tried to make individualplant completely independent so we could reuse for modals
