--- gatsby-starter-default/.prettierignore
@@ -1,4 +1,0 @@
-.cache
-package.json
-package-lock.json
-public
