--- gatsby-starter-default/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "arrowParens": "avoid",
-  "semi": false
-}
