+++ src/components/IndividualPlant.js
@@ -8,50 +8,31 @@ import { Card } from "react-bootstrap";
 import ListGroupItem from "react-bootstrap/esm/ListGroupItem";
 import ListGroup from 'react-bootstrap/ListGroup';
 const IndividualPlant = (props)=> {
-    const cardInfo =[
-        {id:0,image:"", title:"Sally", name:"Ficus lyrata",family:"Fig",size:"12-15m",soil:"Well draining and peat-based",sun:"Around 5 hours a day",hardiness:"Zones 9-11",water:"Once a week"},
-        {id:1,image:"../assets/plantOne.jpg", title:"Jimmy the Plant", name:"Monestera Deliciosa",family:"Arcaceae",size:"10'-15'",soil:"Potting soil with great drainage",sun:"5-8 hours per day of direct sunlight",hardiness:" Zones 10-12",water:"Water once every 1-2 weeks"},
-        {id:2,image:"../assets/plantOne.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"},
-        {id:3,image:"../assets/plantOne.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"},
-        {id:4,image:"../assets/plantFour.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"},
-        {id:5,image:"../assets/plantThree.jpg", title:"Sally", name:"Ficus lyrata",family:"Fig",size:"12-15m",soil:"Well draining and peat-based",sun:"Around 5 hours a day",hardiness:"Zones 9-11",water:"Once a week"},
-        {id:6,image:"../assets/plantOne.jpg", title:"Jimmy the Plant", name:"Monestera Deliciosa",family:"Arcaceae",size:"10'-15'",soil:"Potting soil with great drainage",sun:"5-8 hours per day of direct sunlight",hardiness:" Zones 10-12",water:"Water once every 1-2 weeks"},
-        {id:7,image:"../assets/plantFour.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"},
-        {id:8,image:"../assets/plantFour.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"},
-        {id:9,image:"../assets/plantFour.jpg", title:"Gwendolynn", name:"Dracaena trifasciata",family:"Asparagaceae",size:"80-100cm",soil:"Free draining potting mix",sun:"Indirect sunlight",hardiness:"Zones 10+",water:"Once every two weeks"}
-
-    ];
+    
     // const renderCard=(card,id) => {
-        return(
-            
-             <div className="IndividualPlant container-fluid d-flex justify-content-center ">
-                    <div className="row">
-                    {cardInfo.map((index)=>
-                    //going to need to make this its own component so that when I call this into the modal it doesn't contain the map
+        return(//going to need to make this its own component so that when I call this into the modal it doesn't contain the map
                     <div className="col-md-4">
                         <section>
                     <div className="IndividualPlant text-left col-sm">
-                        <Card className="IndividualPlant" key={index.id} >
+                        <Card className="IndividualPlant" key={props.id} >
                          
                             <Card.Img variant="top"  src={plantOne} ></Card.Img>
                              <Card.Body >
                             <Card.Title  className="fw-bolder text-center" >
-                             {index.title}
+                             {props.title}
                             </Card.Title>
                             <ListGroup variant="flush" >
-                                <ListGroupItem><span style={{fontWeight:'bold'}}> Name:</span> {index.name}</ListGroupItem>
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Family:</span> {index.family}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}> Name:</span> {props.name}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Family:</span> {props.family}</ListGroupItem>
  
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Size:</span> {index.size}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Size:</span> {props.size}</ListGroupItem>
  
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Soil:</span> {index.soil}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Soil:</span> {props.soil}</ListGroupItem>
  
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Sun:</span> {index.sun}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Sun:</span> {props.sun}</ListGroupItem>
  
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Hardiness:</span> {index.hardiness}</ListGroupItem>
-                                <ListGroupItem><span style={{fontWeight:'bold'}}>Water:</span> {index.water}</ListGroupItem>
- 
- 
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Hardiness:</span> {props.hardiness}</ListGroupItem>
+                                <ListGroupItem><span style={{fontWeight:'bold'}}>Water:</span> {props.water}</ListGroupItem>
                             </ListGroup>
                             <div className="text-end">
                             <button type="button" class="btn btn-secondary btn-sm">Edit Plant</button>
@@ -61,9 +42,8 @@ const IndividualPlant = (props)=> {
                      </div>
                      </section>
                     </div>
-                    )}
-                    </div>
-                </div>
+                    
+
